IoTCloudServices/microservices/routes_microservice/code/requirements.txt
Flask
Flask-Cors
mysql-connector-python
requests